feat: Add open-between-gh-and-ghp.js bookmarklet
This commit adds the open-between-gh-and-ghp.js bookmarklet, which allows users to easily switch between GitHub repository pages and GitHub Pages. The bookmarklet is a JavaScript function that can be executed in the browser's address bar. It detects whether the current page is a GitHub repository or a GitHub Pages site and opens the corresponding page in the other format.

Ref: https://neos21.github.io/bookmarklets/src/open-between-gh-and-ghp.js