feat: Add shorten-amazon-url.js bookmarklet
This commit adds the shorten-amazon-url.js bookmarklet, which allows users to shorten Amazon URLs. The bookmarklet is a JavaScript function that can be executed in the browser's address bar. It retrieves the ASIN (Amazon Standard Identification Number) from the page and redirects the user to the shortened URL on Amazon.jp.

Ref: shorten-amazon-url.js